third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Invalid suborigin names</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
var test_suborigin_names = [
  '',
  '\'foobar\'',
  '\'foobar',
  'foobar\'',
  'foo\'bar',
  'foob@r',
  'foo bar',
  'Foobar',
  'FOOBAR',
  '42',
  'foo-bar',
  '-foobar',
  'foobar-'
];

async_test(t => {
    var i = 0;

    window.onmessage = t.step_func(function(event) {
        assert_equals(event.data, 'I am a secret');
        next();
      });

    function next() {
      if (i >= test_suborigin_names.length) {
        t.done();
        return;
      }

      var iframe = document.createElement('iframe');
      iframe.src =
        'resources/reach-into-iframe.php?childsuborigin=' +
        test_suborigin_names[i];
      document.body.appendChild(iframe);

      i++;
    }

    next();
  }, 'Invalid suborigin names');
</script>
</body>
</html>
